GameplayController updated and test
Had to make minor changes to GameplayController class to make sure
everything functioned correctly and that tests passed and acted as they
should.
Made adjustments to RegisterControllerTest and updated the inventory.csv
to include one item on the user at the beginning of the game so that I
could effectively test the user dropping an item.